+++ HISTORY.md
@@ -1,5 +1,9 @@
 # HISTORY
 
+## 17/1/19
+- DMD Ripping: show how many DMD frames has been ripped
+- DMD Ripping: increased maximal ripped DMD Frames from 1000 to 8000
+
 ## 16/1/19
 - add game "Strike Master Shuffle Alley"
 - add game "League Champ Shuffle Alley"